requirements.txt
arrow==1.3.0
binaryornot==0.4.4
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.5.0
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.1
Flask-Login==0.6.3
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
Pygments==2.17.2
python-dateutil==2.8.2
python-slugify==8.0.3
python-webpack-boilerplate==1.0.2
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
urllib3==2.2.0
Werkzeug==3.0.1
WTForms==3.1.2
